test v1

+++ contracts/LoanMaker.sol
@@ -18,14 +18,14 @@ contract LoanMaker {
     }
 
     function createLoan(
-        uint256 _poolSupplyMax,
-        uint256 _numberBorrowerTokens,
-        uint256 _reserveFactorMantissa,
+        uint _poolSupplyMax,
+        uint _numberBorrowerTokens,
+        uint _reserveFactorMantissa,
         address _borrower,
         address _borrowerToken,
         address _admin
-    ) external isadmin returns (uint256) {
-        uint256 _id = allLoans.length;
+    ) external isadmin returns (uint) {
+        uint _id = allLoans.length;
 
         togLoan _loan = new togLoan(
             _poolSupplyMax,
